refactor: loads of refactoring, project structure updates
- first hand support for chrome and firefox
