+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 .envrc
 .direnv
 .dist
+.secrets
