--- Taskfile.yml
@@ -1,44 +1,0 @@
-version: 3
-
-env:
-  VITE_BUILD_DIR_FIREFOX: ".dist/firefox"
-  VITE_BUILD_DIR_CHROME: ".dist/chrome"
-
-tasks:
-  firefox:dev:
-    env:
-      VITE_RUN_MODE: dev
-      VITE_BUILD_FOR_FIREFOX: true
-    cmds:
-      - fwatcher --exec "pnpm run build" --ext ".jsx" --ext ".js" --ext ".json" --ext ".tsx" --ext ".ts" --exclude-dir ".dist"
-
-  chrome:dev:
-    cmds:
-      - fwatcher -e "pnpm run build" --ext ".jsx" --ext ".js" --exclude-dir ".dist"
-
-  firefox:build:
-    env:
-      VITE_BUILD_FOR_FIREFOX: "true"
-    cmds:
-      - npm run build
-
-  chrome:build:
-    cmds:
-      - npm run build
-
-  firefox:zip:build:
-    cmds:
-      - task: firefox:build
-      - |+
-        pushd .dist/firefox
-        zip -r ../../control-center.zip *
-        popd
-
-  zip:source:
-    cmds:
-      - |+
-        dir_name=$(basename $PWD)
-        pushd ..
-        zip -r control-center-source.zip $dir_name -x "$dir_name/node_modules/*" -x "$dir_name/.dist/*" -x "$dir_name/.git/*" -x "$dir_name/.idea/*" -x "$dir_name/.direnv/*"
-        mv ./control-center-source.zip ./$dir_name/control-center-source.zip
-        popd
