--- manifest.chrome.json
@@ -1,53 +1,0 @@
-{
-	"manifest_version": 3,
-	"name": "Control Center",
-	"version": "1.3.2",
-
-	"description": "Control Center for Chrome",
-
-	"icons": {
-		"48": "icons/logo-48.png"
-	},
-
-	"content_scripts": [
-		{
-			"matches": ["*://open.spotify.com/*"],
-			"js": ["src/content-scripts/spotify.com.js"]
-		},
-		{
-			"matches": ["*://music.youtube.com/*"],
-			"js": ["src/content-scripts/music.youtube.com.js"]
-		},
-		{
-			"matches": ["*://www.youtube.com/*"],
-			"js": ["src/content-scripts/youtube.com.js"]
-		}
-	],
-
-	"background": {
-		"service_worker": "src/scripts/background.ts"
-	},
-
-	"options_ui": {
-		"page": "src/options.html"
-	},
-
-	"permissions": ["activeTab", "storage", "scripting", "tabs"],
-
-	"action": {
-		"default_icon": {
-			"48": "icons/logo-48.png"
-		},
-		"default_title": "Control Center"
-	},
-
-	"commands": {
-		"control-center": {
-			"suggested_key": {
-				"default": "Alt+J",
-				"mac": "Alt+J"
-			},
-			"description": "toggle control center tab"
-		}
-	}
-}
