+++ src/app/stores/types.ts
@@ -1,0 +1,9 @@
+export interface HLText {
+	text?: string;
+	hl: boolean;
+}
+
+export interface HLMatchMeta {
+	title: HLText[];
+	url: HLText[];
+}
