--- src/background.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport"
-          content="width=device-width, initial-scale=1" />
-    <meta name="theme-color"
-          content="#000000" />
-    <title>Control Center</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root">
-    </div>
-    <script src="/src/main.tsx"
-            type="module"></script>
-  </body>
-</html>
