+++ src/content-scripts/music.youtube.com/player-controls.ts
