--- src/lib/webext-apis/browser-commands.js
@@ -1,16 +1,0 @@
-export const browserCommands = {
-  "Extensions": async () => {
-    console.log("i am here");
-    return browser.tabs.create({
-      url: 'about:addons',
-      active: true,
-    })
-  },
-
-  "Debug Addons": async () => {
-    return browser.tabs.create({
-      url: browser.runtime.getURL('about:debugging#/runtime/this-firefox'),
-      active: true,
-    })
-  },
-};
