+++ src/manifest/chrome/manifest.json
@@ -1,0 +1,54 @@
+{
+	"manifest_version": 3,
+	"name": "Control Center",
+	"version": "1.3.2",
+
+	"description": "Control Center for Chrome",
+
+	"icons": {
+		"48": "icons/logo-48.png"
+	},
+
+	"content_scripts": [
+		{
+			"matches": ["*://open.spotify.com/*"],
+			"js": ["content-scripts/open.spotify.com/player-controls.js"]
+		},
+		{
+			"matches": ["*://music.youtube.com/*"],
+			"js": ["content-scripts/music.youtube.com/player-controls.js"]
+		},
+		{
+			"matches": ["*://www.youtube.com/*"],
+			"js": ["content-scripts/youtube.com/removes-youtube-shorts.js"]
+		}
+	],
+
+	"background": {
+		"service_worker": "background/index.js",
+		"type": "module"
+	},
+
+	"options_ui": {
+		"page": "option/index.html"
+	},
+
+	"permissions": ["activeTab", "storage", "scripting", "tabs"],
+
+	"action": {
+		"default_icon": {
+			"48": "icons/logo-48.png"
+		},
+		"default_title": "Control Center"
+	},
+
+	"commands": {
+		"control-center": {
+			"suggested_key": {
+				"default": "Alt+J",
+				"mac": "Alt+J"
+			},
+			"description": "toggle control center tab"
+		}
+	}
+}
