+++ src/pkg/dom/observe-and-act.js
@@ -1,0 +1,12 @@
+export function observeAndAct(action) {
+	const observer = new MutationObserver(() => {
+		action();
+	});
+
+	observer.observe(document.body, {
+		subtree: true,
+		childList: true,
+	});
+
+	return observer;
+}
