+++ src/pkg/logger/index.ts
@@ -5,44 +5,78 @@ export interface Logger {
 	error: (err: Error, msg?: string) => void;
 }
 
-function handleKV(...kv): string {
-  let result = ""
-  for (let i = 1; i < kv.length; i+=2) {
-    result += `${kv[i-1]}=${kv[i]}`
-  }
+export function newLogger(name?: string): Logger {
+	console.log(`
+ ██████╗ ██████╗ ███╗   ██╗████████╗██████╗  ██████╗ ██╗          ██████╗███████╗███╗   ██╗████████╗███████╗██████╗ 
+██╔════╝██╔═══██╗████╗  ██║╚══██╔══╝██╔══██╗██╔═══██╗██║         ██╔════╝██╔════╝████╗  ██║╚══██╔══╝██╔════╝██╔══██╗
+██║     ██║   ██║██╔██╗ ██║   ██║   ██████╔╝██║   ██║██║         ██║     █████╗  ██╔██╗ ██║   ██║   █████╗  ██████╔╝
+██║     ██║   ██║██║╚██╗██║   ██║   ██╔══██╗██║   ██║██║         ██║     ██╔══╝  ██║╚██╗██║   ██║   ██╔══╝  ██╔══██╗
+╚██████╗╚██████╔╝██║ ╚████║   ██║   ██║  ██║╚██████╔╝███████╗    ╚██████╗███████╗██║ ╚████║   ██║   ███████╗██║  ██║
+ ╚═════╝ ╚═════╝ ╚═╝  ╚═══╝   ╚═╝   ╚═╝  ╚═╝ ╚═════╝ ╚══════╝     ╚═════╝╚══════╝╚═╝  ╚═══╝   ╚═╝   ╚══════╝╚═╝  ╚═╝
 
-  return result
-}
+  - to enable verbose debug logging, set \`globalThis.control_center_debug = true\`
+`);
 
-export function newLogger(name: string): Logger {
+	const prefix = name ? `control-center/${name}` : "control-center";
+
+	if (import.meta.env.VITE_RUN_MODE === "dev" && globalThis != null) {
+		globalThis.control_center_debug = true;
+	}
+
+	const styleMap = {
+		DEBUG:
+			"background: #9CA3AF; color: black; padding: 2px 4px; border-radius: 2px;",
+		INFO: "background: #2563EB; color: white; padding: 2px 4px; border-radius: 2px;",
+		WARN: "background: #F59E0B; color: black; padding: 2px 4px; border-radius: 2px;",
+		ERROR:
+			"background: #DC2626; color: white; padding: 2px 4px; border-radius: 2px;",
+		PREFIX: "color: gray;",
+		RESET: "",
+	};
+
 	return {
 		debug: (msg: string, ...kv) => {
-			if (kv.length >= 2) {
-			  console.debug(`[${name}] ${msg}`, handleKV(...kv));
-			  return
+			if (globalThis.control_center_debug) {
+				console.debug(
+					`%cDEBUG%c [${prefix}]%c ${msg}`,
+					styleMap.DEBUG,
+					styleMap.PREFIX,
+					styleMap.RESET,
+					...kv,
+				);
+				// console.debug(`[${prefix}] ${msg} |`, ...kv);
 			}
-			console.debug(`[${name}] ${msg}`);
 		},
 		info: (msg: string, ...kv) => {
-			if (kv.length >= 2) {
-			  console.info(`[${name}] ${msg}`, handleKV(...kv));
-			  return
-			}
-			console.info(`[${name}] ${msg}`);
+			console.info(
+				`%cINFO%c [${prefix}]%c ${msg}`,
+				styleMap.INFO,
+				styleMap.PREFIX,
+				styleMap.RESET,
+				...kv,
+			);
+			// console.info(`[${prefix}] ${msg} |`, ...kv);
 		},
 		warn: (msg: string, ...kv) => {
-			if (kv.length >= 2) {
-			  console.warn(`[${name}] ${msg}`, handleKV(...kv));
-			  return
-			}
-			console.warn(`[${name}] ${msg}`);
+			console.warn(
+				`%cWARN%c [${prefix}]%c ${msg}`,
+				styleMap.INFO,
+				styleMap.PREFIX,
+				styleMap.RESET,
+				...kv,
+			);
+			// console.warn(`[${prefix}] ${msg} |`, ...kv);
 		},
 		error: (err: Error, msg?: string, ...kv) => {
-			if (kv.length >= 2) {
-			  console.error(`[${name}] ${msg}`, handleKV(...kv));
-			  return
-			}
-			console.error(`[${name}] ${msg}`, { error: err });
+			console.error(
+				`%cERROR%c [${prefix}]%c ${msg}`,
+				styleMap.INFO,
+				styleMap.PREFIX,
+				styleMap.RESET,
+				{ error: err },
+				...kv,
+			);
+			// console.error(`[${prefix}] ${msg} |`, { error: err }, ...kv);
 		},
 	};
 }
